Move yaml to json magic to yamltojson package
